requirements.txt
fastapi
uvicorn
openai
faiss-cpu
numpy>=1.26
pydantic>=2.5
python-dotenv
prometheus-fastapi-instrumentator
pypdf2           # instead of pypdf (3.12 compatible)
python-docx>=1.1.0  # version that supports Python 3.12
ragas
datasets
